fixed error bug - now getting above 85% accuracy

+++ code/task3/MNIST_reader.h
@@ -46,8 +46,7 @@ void loadMnistImages(const string& filename, vector< vector< double > > &images)
 				{
 					unsigned char pixel = 0;
 					file.read((char*)&pixel, sizeof(pixel));
-					// Normalize (((OldValue - OldMin) * NewRange) / OldRange) + NewMin
-					images[i][n_rows * r + c] = (double) pixel / 255; 
+					images[i][n_rows * r + c] = (float) pixel / 255;
 				}
 			}
 		}
